Finished the forgot password part but didn't test
So every reference to email was removed. And the account controller was
updated to take the phone number from the Dto instead of waiting for it
to show up from localStorage.
-It compiles proper but i didn't test to see if the view or the api
calls work.
-forgot password stuff was added and uses account service to managed.
-I feel like the most likely place to get errors is the validation logic
I copied from the site. I either have to write my own stuff soon if it
doesn't already automatically check correctly for itself.