Finished forgot password testing
I managed to get it to work, But the api isn't giving me anything.
I tried logging in to and it doesn't work as it's supposed to. Maybe
Yewo has something to say as to why it's not working. But honest he
should be having the same error cause I got my repo from him.