Reset password has been done but hasn't been tested
Everything was pretty much procedural except we had to borrow some
custom validator code. We made a new folder to put it into called
customValidators. We also updated the forgotpassword.ts to include the
correct url to get to the reset-password page.
-I want to also remind you that it hasn't been tested. We can't get
past a certain point in the forgot password part so it's impossible to
reach reset password at the moment. I will continue to work on another
branch until I talk over the situation with Yewo.