Tagged sensitive places and dark scalfolding
-I was working on making the repo more secure so I tagged all the places
with possible need for sensitive information. The solution is solved
cause we are using different keys and they aren't on source control any
more but I still taged them in case we decide to use environment
variables and don't want to use config. Although most likely we are
still going to and just simply update the config to reflect our new
changes

-I set up angular material for the dark and light mode for the website.
I didn't actually put in the code for it to respond nor the element
color. I set up the packages for
both in case either fails