This is scafolding for the Reform due
-We are considering using Clean architeture in the Rodizio Express
project. To do so, one of the key move is to have our 'Core' app, (ie
the domain logic and application logic) in a separate nuget package that
all solutions will reference. Each solution will then have its own
Infrastructure and Presentation project and test projects that will be
maintained individually.
-This was just the first step. I very much anticipate making this in a
completely different solution file instead of in RodizioExpress so that
it will be easier to refer to and maintain in the future.