A BillingController was made
-So the scaffolding was put up so that a manager or a franchisee user
would be billed every month. The actual logic and work hasn't been fully
done.
-Its worth considering that there might be a different approach to what
we have. As it stands I first imagined the user would log in and then
logic would fire that would check to bill the user. But that doesn't
make sense. In the next session set out to have something automatically
check if the past due dates have arrived. Then it should fire the
Billing controller logic