Billing view scaffolding has been put up
-So the view has started
-There is still little to know communication between the API and the
client. I haven't got the right types and I'm still trying to figure out
what I am trying to show
-In the API changes are starting to be made. As it were we had a
BillingInfoDto that had two properties. I'm starting to think we can get
those two properties from the UserDto and the systems date.
Because we need a Dto to give all the information from the billing
processing back to the view
-I'm also about to start making a firebaseService. We can't be repeating
code all over the place so I'm thinking we use it to get all our
information we need from the database