So I added a firebaseService
-The firebaseService will be incharge of all things firebase to all the
services and controllers in need of it. This and actions like these are
the scaffolding for clean architeture that we can use.
The service you can imagine will be really larger. So I want to have the
class be divided up to partial classes so that it can house different
concerns and we can monitor them much easier. The interface however, I
want to have all the methods, it will be divided only with regions so
that we can see what is happening in all the firebase concerns
-The firebaseService is already injected into the billingController and
it works perfectly, tested and all.
-The firebaseService was needed in the billingService so it was injected
there