So alot happened...
-See I made the mistake of not committing my changes so there are alot
of things going on that its kind of hard to define what was going on.
-Firstly I made a accountManagement component so its easier for the user
to navigate and access some features and functions that are at their
disposal.
-I'm currently working to make it so that the developer can come in and
set a User to a billed user.
-The problem I was encountering is that I would get a response from the
API for branches that exist. But the moment we get out of the scope it
becomes null, or undefined. And I want to show the branches that a user
(not developer) is managing and the cost each branch has on him/her and
when it is due.
-I set the launch and task json so that I can be able to debug in vscode
-I set all the a elements to red and brown when you hover over them
-Included accountManager in the nav dropdown