Work to put invoice into billingService
-Alot was done. I couldn't stop cause alot of it was tied together. And
to be honest I was to brain dead tired to actually make some good steps
in what I was doing. In fact alot of the work done here will need to be
refactored because its all just poor quality code. It hasn't even been
tested. So alot rides on the next commit. But I will attempt to put into
words what I tried to do here

-Installed the Microsoft.Interlop package so that I can read and write
bookmarks on Word documents. I made a .docx that has bookmarks of texts
I want to replace automatically
-The logic for the text replacement is in billingServices. I'm
considering making a WordService because it is alot
-Created a DateTime extension class so change the date from different
formats and even from orderNumbers
-In firebasesContext I realise that Yewo made some Changes to GetData.
Its a generic method there. I comment how that's a bad idea because you
are forced to make useless types. The alternative I provide works
better.
-BilledUserDto was modified to have the sales from the month and the
username. Consider Refactoring to just use the username from the AppUser
-AdminUser has properties added to make a proper invoice to send to. We
also change alot of the doubles we use for prices to floats cause
doubles are just too precise and large for nothing. And the view works
better with doubles.
-SMS class was added but the logic that was used to work with it is the
most upsetting. Please please look through this code once you are done
with all the billingServices required
-Exceptions for different scenarios were anticipated and created. Look
through them to find references
-I finnally made the ReportService. But I didn't add all the methods I
wanted and only used the ones necessary to move forward. I was force to
cause the billingService needed alot of information and it was difficult
getting them through the ReportController
-FirebaseServices has the weird sms method added
-AccountController was changed to cater for the extraction of the
GetOrdersByDate() method into the ReportServices

-- Summary --
The majority of the changes were done in BillingServices, which why I
was so desparate for a WordService to make the code more manageable.
The problems I find apparent in this exhausted state are how we handle
the SMS logic and the view for the Word doc. If those are solved much of
the clutter should leave