Changes learnt from testing
-As it stands it still doesn't work as it should. But the lessons
learnt from the process introduced changes.

-We put firebaseservices in a normal place
-We changed RegisterDto to have the properties needed to have an
adminUser
-The SendBill method was made to work. It wasn't working cause it was
internal instead of public, and the name of the variable past in was
BillingInfo instead of BillingDto.

-Now to describe the problems I am facing.
So the method fires and when it comes to getting accounts or a small
number of orders (eg 72) its fine. But when it is trying to get all 655
of the completed orders it complains saying the index is out of range.
But that same method works perfectly for the smaller numbers, its to
look into that.
The second problem is that when the code hits the CreateInvoice method,
it won't even enter it, it just throws an error. I didn't have the
energy to look into that too but theres something there

I'm leaving things as they are here cause I want to work on the payment
intergration.`