Merge conflicts with firebase
-Both yewo and I had refactored how we use firebase. He used the
datacontext raw and I used the service. I refactored so that we use the
service but whats interesting is that he had a method that worked for
any type. I was going to do that eventually but yewo beat me to it. So
it was nice to use. We need to have a second overload of the same so
that it works with single unit return types so that we can have most of
our methods removed and slim the firebaseServices