Changed how some metrics were displayed
-Some metrics were coming up as wrong and they were misleading and
depressing. So I changed a few (not all to be more closer to the truth)
A warning was given in the previous commit. If you truely feel
compelled to keep the previous version, we can discuss it and revert the
work
-Refactored somemore code. In all honesty there is sooooo much more to
do, ReportController is terrible. I'm just glad I somehow stopped myself
from continuing. Time is wasting