Streamlined Excel and other things
-I first want to note that accountService was given GetAdminAccounts. so
everywhere like Billing and accounts and firebaseService was done to
reflect that change
-Next I started refactoring ExcelController and I quickly found
repeating code that all didn't need to be in a controller. So as I made
and populated the excelService I ended up moving everything there. There
is only one business state that needed the excel work so it was given
that access, ReportController.
-The final change was in the view so that it doesnt call an endpoint
that doesn't exist. The excel endpoint for all the orders in a branch
was moved to the report controller.