Environment and JsonConverter
-Locally environments were introduced so as to switch between prod and
deve variables for the url the client should call to
-The changes were then made to the BaseService to have that happen
-The Jsonconverter was throwing errors cause I was using JObject for
arrays instead of JArray. Once that was fixed the wrong method was being
involved cause the compilter couldn't tell the difference between
List<objects> and an object. So we will only use one of the generic
methods and it works for both instances
-AdminUser had repeating code that it inherited from AppUser (email)
`