Report Controller finnally in sync
-So The Jsonconverter was having trouble with distinguishing between an
aggregate and an entity. For a while I tried to have the generic method
some how cater for both. But I pretty much wasted a day making that
happen only to fail. I eventually stopped being prideful and made to
methods. To it is important that when you are dealing with aggregates
you have to use the DataArray firebase method