Brush through before continuing
-So we are trying to make a payment and invoicing feature for the app.
Yewo seems to have a feature already up and running but I don't know how
much of it is actually automated.

-In any case we made some services use more of appsettings, and we
attempted to make nUnit test project. Without wifi it is difficult to
figure out why the classes aren't working as expected.
-We shifted two methods from billing service to reportService where it
was thought it would be more suited there