Clean Architeture introduced
-So alot of the clean architeture was done with paySandbox because the
services needed work. Alot of work.
So I simply pulled from there and made sure the changes compile. The
problem comes that its was not a lean merge. Alot of useless code still
exists in here and so with the next commit i hope to remove all of it so
I can get something that makes more sense.

-All of the Business Core logic is brought in from the company Kernel.
That way changes are more seemless between solutions
-The files are more appropriately arranged but the namespaces aren't
reflecting yet.